@@ -50,7 +50,6 @@
     "@types/serve-favicon": "^2.5.1",
     "@types/session-file-store": "^1.2.2",
     "applicationinsights": "^2.2.0",
-    "autobind-decorator": "^2.4.0",
     "awilix": "^6.1.0",
     "axios": "^0.25.0",
     "config": "^3.3.7",
@@ -61,15 +60,15 @@
     "express-nunjucks": "^2.2.5",
     "express-session": "^1.17.2",
     "glob": "^7.1.6",
+    "helmet": "^5.0.2",
     "govuk-frontend": "^4.0.0",
-    "helmet": "^4.2.0",
     "jest-when": "^3.5.0",
-    "jquery": "^3.5.1",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.20",
     "npm-run-all": "^4.1.5",
     "nunjucks": "^3.2.2",
     "redis": "^3.1.2",
+    "request": "^2.88.2",
     "require-directory": "^2.1.1",
     "serve-favicon": "^2.5.0",
     "session-file-store": "^1.5.0",
@@ -81,11 +80,9 @@
     "@babel/core": "^7.16.12",
     "@babel/preset-env": "^7.16.11",
     "@codeceptjs/helper": "^1.0.2",
-    "@types/chai-as-promised": "^7.1.1",
     "@types/copy-webpack-plugin": "^10.1.0",
     "@types/jest": "^27.4.0",
     "@types/mini-css-extract-plugin": "^2.5.1",
-    "@types/supertest": "^2.0.8",
     "@types/webpack-dev-middleware": "^5.3.0",
     "@typescript-eslint/eslint-plugin": "^2.19.2",
     "@typescript-eslint/eslint-plugin-tslint": "^5.10.2",
@@ -93,35 +90,27 @@
     "allure-commandline": "^2.13.8",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.3",
-    "chai": "^4.3.6",
-    "chai-as-promised": "^7.1.1",
     "codeceptjs": "^3.2.3",
-    "copy-webpack-plugin": "^6.1.0",
-    "css-loader": "^3.4.2",
-    "debug": "^4.3.3",
+    "copy-webpack-plugin": "^10.2.0",
+    "css-loader": "^6.5.1",
     "eslint": "^6.8.0",
     "eslint-plugin-codeceptjs": "^1.3.0",
-    "html-webpack-plugin": "^4.3.0",
-    "jest": "^25.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^27.4.7",
     "jest-html-reporter": "^3.4.2",
-    "mini-css-extract-plugin": "^1.6.2",
+    "mini-css-extract-plugin": "^2.5.2",
     "nock": "^13.2.2",
-    "node-sass": "^6.0.1",
+    "node-sass": "^7.0.1",
     "nodemon": "^2.0.15",
     "pa11y": "^6.1.1",
     "playwright": "^1.18.1",
-    "request": "^2.88.0",
-    "sass-loader": "^10.0.2",
-    "sinon": "^13.0.1",
-    "sinon-chai": "^3.4.0",
+    "sass-loader": "^12.4.0",
     "sonar-scanner": "^3.1.0",
-    "style-loader": "^2.0.0",
-    "supertest": "^6.2.2",
-    "ts-jest": "^25.2.1",
+    "style-loader": "^3.3.1",
+    "ts-jest": "^27.1.3",
     "ts-loader": "^9.2.1",
     "tslint": "6.1.3",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "webpack": "^4.41.6",
+    "webpack": "^5.66.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-middleware": "^5.3.1",
     "webpack-node-externals": "^3.0.0"
